Interim

+++ app/src/main/java/de/dh/apstest/data/db/entities/DBTargetBlock.kt
@@ -1,3 +1,3 @@
 package de.dh.apstest.data.db.entities
 
-data class DBTargetBlock(val duration: Short, val lowTarget: Int, val highTarget: Int)
+data class DBTargetBlock(val duration: Short, val lowTarget: Short, val highTarget: Short)
